Start working on interpreting STL containers in TTrees.
